// src/app/models/user.model.ts
export class UserModel {
    id?: string;
    chatId?: string;
    email?: string;
    name?:string;
    plants?:string[];
    surnames?:string;
    telephone?:string;
    max_air_humidity?: number;
    min_air_humidity?: number;
    max_soil_humidity?: number;
    min_soil_humidity?: number;
    max_temperature?: number;
    min_temperature?: number;
    
  }